Refactored src into sub packages
