+++ src/db/db_config.py
@@ -1,4 +1,4 @@
-from .app import app
+from ..app import app
 from flaskext.mysql import MySQL
 
 mysql = MySQL()
