+++ src/service/response_templates.py
